backend/requirements.txt
# Smart TCO Calculator - Backend Dependencies

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data processing
pandas==2.2.0
numpy==1.26.3
scikit-learn>=1.3.0  # For Random Forest and t-SNE visualization

# Google Cloud / Vertex AI (optional)
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.3.0
vertexai>=1.60.0

# Embeddings (optional, for RAG)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# PDF processing for RAG knowledge base
pypdf2>=3.0.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0  # For API calls (ENTSO-E energy prices)

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
